pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screen-record"
version = "0.1.0"
description = "Cross-platform screen recorder with editable keystroke overlays."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.8",
  "pynput>=1.7.7",
  "mss>=9.0",
  "Pillow>=10.0",
]

[project.scripts]
screen-record = "screen_record.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pyinstaller>=6.19.0",
]
